auto-video/requirements.txt
ffmpeg_python==0.2.0
google_api_python_client==2.96.0
google_auth_oauthlib==1.0.0
openai==0.27.8
